Refactor database-related dependencies: done for GenericFilter and Settings
