+++ jmix-bom/bom.gradle
@@ -162,6 +162,8 @@ dependencies {
         api "io.jmix.flowui:jmix-flowui-themes:$freeVersion"
         api "io.jmix.flowui:jmix-flowui-data:$freeVersion"
         api "io.jmix.flowui:jmix-flowui-data-starter:$freeVersion"
+        api "io.jmix.flowui:jmix-flowui-restds:$freeVersion"
+        api "io.jmix.flowui:jmix-flowui-restds-starter:$freeVersion"
         api "io.jmix.flowui:jmix-flowui-devserver:$freeVersion"
         api "io.jmix.flowui:jmix-flowui-test-assist:$freeVersion"
