+++ jmix-flowui/flowui-data/src/main/resources/io/jmix/flowuidata/messages.properties
@@ -36,26 +36,6 @@ io.jmix.flowuidata.entity/UserSettingsItem.username=Username
 io.jmix.flowuidata.entity/UserSettingsItem.createTs=Created at
 io.jmix.flowuidata.entity/UserSettingsItem.createdBy=Created by
 
-actions.GenericFilter.MakeDefault = Make default
-actions.GenericFilter.Save = Save
-actions.GenericFilter.SaveAs = Save as
-actions.GenericFilter.SaveWithValues = Save with values
-actions.GenericFilter.Remove = Remove
-
-io.jmix.flowuidata.action.genericfilter/genericFilterRemoveAction.confirmationDialog.header = Confirmation
-io.jmix.flowuidata.action.genericfilter/genericFilterRemoveAction.confirmationDialog.text = Are you sure you want to \
-  delete the current configuration?
-
-io.jmix.flowuidata.action.genericfilter/saveFilterConfigurationInputDialog.header = Save filter configuration
-io.jmix.flowuidata.action.genericfilter/saveFilterConfigurationInputDialog.idField.emptyValue = Id field should be filled
-
-io.jmix.flowuidata.component.genericfilter.configuration/availableForAllUsersField.label = Available for all users
-io.jmix.flowuidata.component.genericfilter.configuration/generatedIdField.label = Generated id
-io.jmix.flowuidata.component.genericfilter.configuration/nameField.nonUniqueUserName = User configuration with same \
-  name already exists
-io.jmix.flowuidata.component.genericfilter.configuration/nameField.nonUniqueGlobalName = Global configuration with \
-  same name already exists
-
 io.jmix.flowuidata.exception/optimisticLock.exceptionTitle = Unable to save the '%s' object changes
 io.jmix.flowuidata.exception/optimisticLockUnknownObject.message = Unable to save the object changes
 io.jmix.flowuidata.exception/optimisticLock.exceptionMessage = The object has just been changed by another user or by the system. \n\
