+++ jmix-flowui/flowui-restds-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.jmix.autoconfigure.flowuirestds.FlowuiRestDsAutoConfiguration
