+++ jmix-flowui/flowui-restds/src/main/java/io/jmix/flowuirestds/FlowuiRestDsConfiguration.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2022 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.jmix.flowuirestds;
+
+import io.jmix.core.annotation.JmixModule;
+import io.jmix.flowui.FlowuiConfiguration;
+import io.jmix.restds.RestDsConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan
+@JmixModule(dependsOn = {RestDsConfiguration.class, FlowuiConfiguration.class})
+public class FlowuiRestDsConfiguration {
+}
