+++ jmix-flowui/flowui/src/main/java/io/jmix/flowui/accesscontext/UiGenericFilterModifyGlobalConfigurationContext.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2023 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.jmix.flowui.accesscontext;
+
+import io.jmix.core.accesscontext.SpecificOperationAccessContext;
+
+/**
+ * Specific operation access context to check whether a user has permission to modify global filter configurations.
+ */
+public class UiGenericFilterModifyGlobalConfigurationContext extends SpecificOperationAccessContext {
+
+    public static final String NAME = "ui.genericfilter.modifyGlobalConfiguration";
+
+    public UiGenericFilterModifyGlobalConfigurationContext() {
+        super(NAME);
+    }
+}
