+++ jmix-flowui/flowui/src/main/java/io/jmix/flowui/component/genericfilter/FilterConfigurationPersistence.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2024 Haulmont.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.jmix.flowui.component.genericfilter;
+
+import io.jmix.flowui.component.genericfilter.model.FilterConfigurationModel;
+import org.springframework.lang.Nullable;
+
+import java.util.List;
+
+public interface FilterConfigurationPersistence {
+
+    void remove(FilterConfigurationModel configurationModel);
+
+    void save(FilterConfigurationModel configurationModel);
+
+    @Nullable
+    FilterConfigurationModel load(String configurationId, String componentId, String username);
+
+    List<FilterConfigurationModel> load(String componentId, String username);
+}
