+++ jmix-flowui/flowui/src/main/resources/io/jmix/flowui/messages.properties
@@ -47,6 +47,12 @@ actions.genericFilter.Copy=Copy
 actions.genericFilter.Edit=Edit
 actions.genericFilter.Reset=<Reset filter>
 
+actions.GenericFilter.MakeDefault = Make default
+actions.GenericFilter.Save = Save
+actions.GenericFilter.SaveAs = Save as
+actions.GenericFilter.SaveWithValues = Save with values
+actions.GenericFilter.Remove = Remove
+
 boolean.yes=Yes
 boolean.no=No
 
@@ -402,6 +408,20 @@ io.jmix.flowui.component.genericfilter.configuration/filterConfigurationDetail.g
 io.jmix.flowui.action.genericfilter/genericFilterEditAction.filterConfigurationDetail.title = Configuration
 io.jmix.flowui.action.genericfilter/genericFilterEditAction.filterConfigurationDetail.groupConditionTitle = Root group condition
 
+io.jmix.flowui.action.genericfilter/genericFilterRemoveAction.confirmationDialog.header = Confirmation
+io.jmix.flowui.action.genericfilter/genericFilterRemoveAction.confirmationDialog.text = Are you sure you want to \
+  delete the current configuration?
+
+io.jmix.flowui.action.genericfilter/saveFilterConfigurationInputDialog.header = Save filter configuration
+io.jmix.flowui.action.genericfilter/saveFilterConfigurationInputDialog.idField.emptyValue = Id field should be filled
+
+io.jmix.flowui.component.genericfilter.configuration/availableForAllUsersField.label = Available for all users
+io.jmix.flowui.component.genericfilter.configuration/generatedIdField.label = Generated id
+io.jmix.flowui.component.genericfilter.configuration/nameField.nonUniqueUserName = User configuration with same \
+  name already exists
+io.jmix.flowui.component.genericfilter.configuration/nameField.nonUniqueGlobalName = Global configuration with \
+  same name already exists
+
 #JmixGridColumnVisibility
 io.jmix.flowui.component.gridcolumnvisibility/showAllItem.text=Show all
 io.jmix.flowui.component.gridcolumnvisibility/hideAllItem.text=Hide all
@@ -427,3 +447,14 @@ richTextEditor.i18n.link=Link
 richTextEditor.i18n.blockquote=Blockquote
 richTextEditor.i18n.codeBlock=Code block
 richTextEditor.i18n.clean=Clean formatting
+
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel = Configuration
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.id = Id
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.componentId = Component id
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.configurationId = Configuration id
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.name = Configuration name
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.username = Username
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.defaultForAll = Default for all users
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.rootCondition = Root condition
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.sysTenantId = Tenant id
+io.jmix.flowui.component.genericfilter.model/FilterConfigurationModel.defaultForMe = Default for me
