+++ settings.gradle
@@ -73,6 +73,8 @@ includeProject('flowui-kit', 'jmix-flowui/flowui-kit')
 includeProject('flowui-themes', 'jmix-flowui/flowui-themes')
 includeProject("flowui-data", 'jmix-flowui/flowui-data')
 includeProject("flowui-data-starter", 'jmix-flowui/flowui-data-starter')
+includeProject("flowui-restds", 'jmix-flowui/flowui-restds')
+includeProject("flowui-restds-starter", 'jmix-flowui/flowui-restds-starter')
 includeProject("flowui-devserver", 'jmix-flowui/flowui-devserver')
 includeProject('flowui-test-assist', 'jmix-flowui/flowui-test-assist')
